requirements.txt
aiohttp==3.9.5 
annotated-types==0.7.0 
anyio==4.4.0 
attrs==23.2.0 
certifi==2024.6.2 
charset-normalizer==3.3.2 
colorama==0.4.6 
dataclasses-json==0.6.7 
distro==1.9.0 
faiss-cpu==1.8.0.post1 
greenlet==3.0.3 
httpx==0.27.0 
langchain==0.2.7 
langchain-community==0.2.7 
langchain-core==0.2.18 
langchain-openai==0.1.16 
langsmith==0.1.80 
marshmallow==3.21.3 
numpy==1.26.4 
openai==1.34.0 
orjson==3.10.5 
pydantic==2.7.4 
PyMuPDF==1.24.5 
pypdf==4.3.0 
PyYAML==6.0.1 
regex==2024.5.15 
requests==2.32.3 
SQLAlchemy==2.0.31 
tenacity==8.4.1 
tiktoken==0.7.0 
tqdm==4.66.4
flask
cryptography>=3.1
load_dotenv
CORS
flask-cors